requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.12
transformers>=4.44.0
torch>=2.4.0
torchvision>=0.19.0
Pillow>=10.4.0
opencv-python-headless>=4.10.0
numpy>=1.26.0
aiofiles>=24.1.0
accelerate>=0.33.0
